refactor: Create diffArchive module
