+++ packages/cli/src/diffArchive/SourceDiff.ts
@@ -2,26 +2,26 @@ import parseDiff from 'parse-diff';
 import { warn } from 'console';
 import assert from 'assert';
 
-import { executeCommand } from '../../lib/executeCommand';
+import { executeCommand } from '../lib/executeCommand';
 import { QueueObject, queue } from 'async';
 import { existsSync } from 'fs';
 
 export const MAX_DIFF_LENGTH = 1000 * 1000;
 
-export type Diff = {
+export type SourceDiffItem = {
   command: string;
   files: parseDiff.File[];
 };
 
-export class DiffLoaderQueue {
+export class SourceDiffQueue {
   private diffQueue: QueueObject<string>;
-  private diffByRoots = new Map<string, Diff>();
+  private diffByRoots = new Map<string, SourceDiffItem>();
 
   constructor(public baseRevision: string, public headRevision: string) {
     this.diffQueue = queue(this.executeDiffCommand.bind(this), 1);
   }
 
-  async diff(sourcePathRoots: Set<string>): Promise<Diff | undefined> {
+  async diff(sourcePathRoots: Set<string>): Promise<SourceDiffItem | undefined> {
     // Process only one diff at a time.
 
     if (this.diffQueue.length()) await this.diffQueue.drain();
@@ -40,7 +40,7 @@ export class DiffLoaderQueue {
     if (this.diffByRoots.get(roots)) return;
 
     const command = `git diff ${this.baseRevision}..${this.headRevision} -- ${roots}`;
-    let diff: Diff = { command, files: [] };
+    let diff: SourceDiffItem = { command, files: [] };
     try {
       const diffStr = await executeCommand(command, true);
       if (diffStr.length > MAX_DIFF_LENGTH) {
@@ -58,13 +58,13 @@ export class DiffLoaderQueue {
   }
 }
 
-export default class DiffLoader {
+export default class SourceDiff {
   private sourcePathRoots = new Set<string>();
-  private diffByFile: Map<string, Diff> | undefined;
-  private diffLoaderQueue: DiffLoaderQueue;
+  private diffByFile: Map<string, SourceDiffItem> | undefined;
+  private diffLoaderQueue: SourceDiffQueue;
 
   constructor(public baseRevision: string, public headRevision: string) {
-    this.diffLoaderQueue = new DiffLoaderQueue(baseRevision, headRevision);
+    this.diffLoaderQueue = new SourceDiffQueue(baseRevision, headRevision);
   }
 
   lookupDiff(path: string): string | undefined {
@@ -94,7 +94,7 @@ export default class DiffLoader {
     for (const root of sourcePathRoots) {
       if (!this.sourcePathRoots.has(root)) {
         this.sourcePathRoots.add(root);
-        if (DiffLoader.isEligibleFile(root)) diffRoots.add(root);
+        if (SourceDiff.isEligibleFile(root)) diffRoots.add(root);
       }
     }
 
@@ -106,8 +106,8 @@ export default class DiffLoader {
     return !['vendor', 'node_modules'].includes(root) && existsSync(root);
   }
 
-  private parseDiffStr(diff: Diff) {
-    if (!this.diffByFile) this.diffByFile = new Map<string, Diff>();
+  private parseDiffStr(diff: SourceDiffItem) {
+    if (!this.diffByFile) this.diffByFile = new Map<string, SourceDiffItem>();
 
     diff.files.reduce((memo, diffFile) => {
       [...new Set([diffFile.to, diffFile.from])]
