+++ packages/cli/src/lib/normalizeAppMapId.ts
