+++ packages/cli/tests/unit/archive/fixtureData/addDeleteChangeRemoveDiff.txt
