+++ packages/cli/tests/unit/archive/fixtureData/exampleDiff.txt
