+++ packages/cli/tests/unit/archive/fixtureData/expectedDiff.txt
