+++ packages/cli/tests/unit/compare/isChanged.spec.ts
@@ -1,7 +1,7 @@
 import { verbose } from '../../../src/utils';
 import { isChanged } from '../../../src/cmds/compare/ChangeReporter';
-import { Digests } from '../../../src/cmds/compare/Digests';
-import { RevisionName } from '../../../src/cmds/compare/RevisionName';
+import { Digests } from '../../../src/diffArchive/Digests';
+import { RevisionName } from '../../../src/diffArchive/RevisionName';
 import { AppMapName } from '../../../src/cmds/compare/ChangeReport';
 
 if (process.env.VERBOSE) verbose(true);
